+++ CHANGELOG.md
@@ -51,6 +51,12 @@ The versions follow [semantic versioning](https://semver.org).
 
   - Apache Velocity Template (Extensions: `.vm`, `.vtl`)
 
+- Some special endings are always stripped from copyright and licensing
+  statements (#602):
+
+  - `">` (and variations such as `'>`, `" >`, and `"/>`)
+  - `] ::`
+
 ### Changed
 
 - Updated PyPI development status to 'production/stable' (#381)